VectorSearch/pyproject.toml
[project]
name = "vectorsearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "sentence-transformers>=3.1.0",
]
